Delete broken TO API test (#3864)
This test currently does not even compile. Even if it did compile, the
TO API test framework does not currently support testing endpoints that
use the to_extension table. This is because the test deletes all rows
from the to_extension table when it starts up, but the to_extension
table is only populated by seeds.sql (which should be done before the
test is run).