Add atscfg error log for multiple cacheurl/key
Adds error logging whenever the config gen inserts multiple cacheurl
or cachekey plugins on the same line.

This isn't supported in ATS, and will usually do the wrong thing.
But, we've always generated config like this, so for now, this
doesn't change the config gen, but it logs the error