Teach dataset loader how to automatically differentiate cache profiles by hardware
This adds support for mixed cache hardware that doesn't use the same kinds of hardware.  It's still a round-robin based association, but the hardware variances are excluded.  This also necessitated a refactor in how profiles were mapped to components.  Instead of using whitelist+blacklist, it's now just a regex match.