Fix CIAB goose, login, and config gen issues (#3956)
* Fix up Docker TO API tests config gen

Fix the code that checks if required env vars are set and properly
return an error with all the unset vars. Fix it to use TO_ADMIN_PASSWORD
(from variables.env) instead of ADMIN_PASS which doesn't actually get
set anywhere.

* Fix CIAB goose and login issues

The goose binary was not being installed properly, and the psql command
to add the TO admin user into the db was failing silently because of
this. Make sure the enroller gracefully exits if it's unable to login
properly.