Remove commented out PATCH /jobs code, api_capability, and docs (#3985)
The PATCH method was not included and should not be referenced by
documentation or exist as an api_capability. In addition, commented-out
code in the codebase is unnecessary cruft that makes the codebase harder
to search and understand. If a PATCH method needs to be supported in the
future, the commented-out code can be retrieved from the git history
instead.