Api test gardening (#4134)
* Remove redundant formatting flags

* Remove unnecessary newlines at end of test failure messages

* Use basic string methods when no formatting verbs present

* Remove trailing newlines from error messages

* Remove unneccesary trailing newline from t.Errorf and t.Fatalf calls

* Exit with code 1 when test suite fails to initialize

The `go test` command inspects the exit code of the test binary it
it builds and runs to determine overall success. Calling os.Exit(0)
or returning from TestMain without calling os.Exit tells `go test` that
the tests have succeeded. It is misleading to do so without having run
any of the tests.

* Allow `go test -list=.` to work