Pin Go builds to 1.11.13
TO was recently pinned to 1.11 due to issues communicating with Riak at
the Go version provided by EPEL (1.13 currently). Since 1.13 isn't
really "vetted" yet, pin the rest of the Go components to 1.11 as well.
Builds should be deterministic, and advancing Go versions should be a
very intentional process with thorough testing.