Fix atstccfg caching
Fixes atstccfg caching, which was breaking when the new cache file
was smaller than the old, because it wasn't truncating the file.

Also changes deliveryservice_server to always fetch all Servers and
Delivery Services and cache them, because it's faster for an ORT
run to reuse and cache them all.