Fix TO Perl endpoints to always be numbers. (#4268)
Fixes the serverchecks and to_extensions endpoints to always return
numbers, as they should.

Apparently, certain versions of Perl or its libraries sometimes turn
the numbers into strings. This forces it to always be a number.

These two endpoints are specifically used in the API tests. This
makes the API tests pass correctly.

We should probably make all numbers in all endpoints do this; but
because Perl is going away, that may not be worth it. But these two
at least let devs verify all API tests pass, instead of potentially
having to ignore failures and hope their code changes didn't cause
other failures in the same tests.