Change atstccfg to use gob,csv caching
Changes atstccfg caching to use gob by default, with a custom csv
format for delivery service servers which is huge and slow.

It also abstracts the cache reading and writing, so it's easy
to change to another format.

It also leaves in a const to easily change it to JSON at compile
time.

On a large ORT run, this cuts the time down to about 1/3 what
it was with JSON.