WIP change atstccfg to generate all files at once
Changes atstccfg to build the meta config, and then build all configs
from it, and output them to a specified directory, in a single run.

This is much faster than independent runs called by ORT, and also
a step toward replacing ORT with atstccfg.

This changeset does not modify ORT as necessary to handle the single,
run, and is thus WIP and not complete or functional yet.