Format DNSSEC keys effectiveDate in RCF3339 format
Currently, TP is formatting the effectiveDate field in the request as
unix epoch in MILLISECONDS. TO API expects epoch time format to be in
SECONDS. This means that TO creates DNSSEC keys with an effective date
ridiculously far into the future, when it was meant for the effective
date to be "now".

Alternatively, just format it as RFC3339, which the TO API can
also parse.