Fix lib/go-tc missing symbol from 4.0.x.
Adds back a symbol which was in 4.0.x. This allows users to vendor
libraries that use lib/go-tc, namely traffic_ops/client.

This is critically important. Without maintaining backward
compatibility in lib/go-tc (which Go Modules will require anyway),
it's impossible to vendor the client, and thus (since we keep the
client in master matching the latest TO Server) impossible to write a
new app using the client to request a production Traffic Ops which
isn't the version of the very latest master branch.

Unless that app also vendored lib/go-tc. But then the Go symbols
of the vendored and unvendored tc don't match and won't compile,
unless the app itself also vendors lib/go-tc. Except any external
libraries it uses which also use lib/go-tc (for example,
lib/go-atscfg) _can't_ be made to use the vendored version. Making
it actually impossible to decompose your app into libraries.

Moreover, writing functions to translate between vendored and
unvendored symbols is also impossible, because Go won't let you
import or declare variables of a vendored type, making it actually
impossible to declare a function with a vendored symbol as the input
or output.

It is theoretically possible to make the code compile, but it is
extremely, extremely painful. I am speaking from experience.