Change ORT/atstccfg to generate all files at once, massive performance improvement (#4493)
* WIP change atstccfg to generate all files at once

Changes atstccfg to build the meta config, and then build all configs
from it, and output them to a specified directory, in a single run.

This is much faster than independent runs called by ORT, and also
a step toward replacing ORT with atstccfg.

This changeset does not modify ORT as necessary to handle the single,
run, and is thus WIP and not complete or functional yet.

* WIP Change atstccfg to geneate multipart file

* WIP add atstccfg args for TO data that ORT needs

* Remove ORT unused api_in_use conditionals

* Remove ORT unused funcs

* Add ORT atstccfg reval-only flag.

Adds a flag to the atstccfg config generator, to only get files
named regex_revalidate.config.

This is needed by ORT, which has a mode to only get reval files.

* Change ORT to use atstccfg single-file generation

* Remove ORT atstccfg unused func, rename func

Removes unused funcs getting data from TO (which were previously used
with the one-at-a-time generation, but not needed by the new
single-file gen).

Also renamed GetCached to GetRetry, to reflect what it actually does
now.

* Fix lib/go-tc missing symbol from 4.0.x.

Adds back a symbol which was in 4.0.x. This allows users to vendor
libraries that use lib/go-tc, namely traffic_ops/client.

This is critically important. Without maintaining backward
compatibility in lib/go-tc (which Go Modules will require anyway),
it's impossible to vendor the client, and thus (since we keep the
client in master matching the latest TO Server) impossible to write a
new app using the client to request a production Traffic Ops which
isn't the version of the very latest master branch.

Unless that app also vendored lib/go-tc. But then the Go symbols
of the vendored and unvendored tc don't match and won't compile,
unless the app itself also vendors lib/go-tc. Except any external
libraries it uses which also use lib/go-tc (for example,
lib/go-atscfg) _can't_ be made to use the vendored version. Making
it actually impossible to decompose your app into libraries.

Moreover, writing functions to translate between vendored and
unvendored symbols is also impossible, because Go won't let you
import or declare variables of a vendored type, making it actually
impossible to declare a function with a vendored symbol as the input
or output.

It is theoretically possible to make the code compile, but it is
extremely, extremely painful. I am speaking from experience.

* Vendor ORT traffic_ops/client

ORT is required to work with at least one version back, so users
aren't required to update in a particular order.

This is necesary, because the master client requests endpoints
which don't exist in the previous TC version.

* Fix intermittent test failure

The unit test was verifying a port 42 wasn't in a config file.
But the timestamp is in the header comment, so whenever the time
contained '42' the test failed.

This changes the port to be a bigger number, so it's less likely.

* Fix broken test, remove unused files

Removes a unit test for functions that no longer exist.

Also moves the last non-caching func from caching.go and removes the
file. Since atstccfg no longer does caching.

* Add ORT atstccfg preprocessing

* Remove unnecessary type declaration

* Fix ORT erroneous return preprocessing

* Remove comment

* Remove ORT atstccfg unused return val

* Fix ORT atstccfg debug logging to info

atstccfg doesn't use the debug logger and it's always nil. Fixed
uses to Info.

* Remove ORT atstccfg unused OutputDir config

* Remove ORT atstccfg PrintGeneratedFiles flag

Flag no longer makes sense, as atstccfg generates all files now.

Some files are dynamically generated, so callers should only have
used the flag to determine if a file was generated by atstccfg or
Traffic Ops, not whether the file exists.

* Fix ORT atstccfg Cfg struct to be ordered

* Fix ORT atstccfg wrong usage string

* Fix ORT atstccfg misnamed log msg

* Add ORT atstccfg comment

* Remove ORT unused imports, variables

* Fix ORT spacing

* Add ORT readme

* Add API Test for TO client GetUpdateStatus

* Add ORT atstccfg unit tests

* Fix ORT atstccfg missing licenses

* Add ORT atstccfg sslkeys file

* Fix ORT to ensure files have exactly 1 \n$

Fixes ORT to trim all trailing whitespace from files, and add exactly
one trailing newline as required by POSIX.

This specifically fixes it to trim carriage returns left in by
the mixed/multipart message.