Add ORT atstccfg TO latest version client
Adds a client of the latest Traffic Ops version, and refactors the
existing vendored client, so both can be used at once.

Includes an idiom of using the latest and falling back to vendored.

Also includes an example of the latest client getting Delivery
Services, and falling back to the vendored if the TO isn't the latest.
This isn't actually necessary right now, but it's likely to be very
shortly, and it serves as an example of the idiom for future dev.