Fix some CIAB issues and add TC_LOC type to seeds.sql
Recently, creating types other than for use in the servers table was
prohibited, so add a TC_LOC type to seeds.sql to be used for cachegroups
containing servers that don't match the existing set of cachegroup
types.

Remove the TC_LOC and TR_LOC types from being created by the CIAB
enroller and fix a few other bugs.