Fix /users/register DB query
The query was failing with the following:

    missing destination name rolename in *tc.User

This indicates that the struct was missing a "db:rolename" tag for that
field, but simply adding that tag would break the /users API due to it
using a special struct for GetUsers.

Instead, convert from sqlx.StructScan to plain sql.Scan, which avoids
any unintended side-effects caused by changing the User struct tags.

Fixes #4583.