Add readiness container to ciab
Today, there's not really a great way to tell when ciab is "ready", i.e.
all services have started up successfully, so that you can start using
it. Or, maybe you only care to know whether ciab started up successfully
or not, and that's it.

Traditionally, ciab being "successful" means that you can curl the
delivery service, so that's what the readiness container does. This way,
you can start up the "main" services in the background and start up the
"readiness" service in the foreground. When ciab is ready, the readiness
service will exit, and you can start using ciab. You no longer manually
have to watch the log output for things that _probably_ indicate ciab
started up successfully, or manually curl the delivery service in a
loop until it responds.