Add Raw Remap Range Directive
Adds a __RANGE_DIRECTIVE__ to the Raw Remap processing, to insert the
Delivery Service Range Request Handling directives there, instead of
before the Raw Remap.

Normally, Raw Remap always needs to be at the end of the remap line,
but in some cases, specifically when custom remaps are modifying
the Range header, it's necessary to put the Range Request directives
(@plugin=slice.so, @plugin=caceh_raange_requests.so, etc) after
the custom Raw Remap text.

This makes that possible.