Updated to weasel v0.4.
This version has a number of feature improvements, including the ability to tag
files directly with SPDX-License-Identifier fields so they are properly
identified even if they move and will recognize a .gitignore file even if the
directory is not a git work tree. This is important for running weasel on
released tarballs.