Add ORT inferring location params (#4861)
Adds ORT location param inference. If location Parameters do not exist
the files are created and added anyway with the directory determined
from the local ATS install, for:
- cache.config
- hosting.config
- ip_allow.config
- parent.config
- plugin.config
- records.config
- remap.config
- storage.config
- volume.config
- Delivery Services with
  - Edge Header Rewrite
  - Mid Header Rewrite,
  - Cache URL
  - URL Sig
  - URI Signing

Note this is not an exhaustive of required files, and many files
must be dynamic and cannot be inferred.

But this does remove the manual configuration for this list.
More files may be added in the future.