Add Topology ATS Config Generation (#4790)
* Add ORT inferring location params

Adds ORT location param inference. If location Parameters do not exist
the files are created and added anyway with the directory determined
from the local ATS install, for:
- cache.config
- hosting.config
- ip_allow.config
- parent.config
- plugin.config
- records.config
- remap.config
- storage.config
- volume.config
- Delivery Services with
  - Edge Header Rewrite
  - Mid Header Rewrite,
  - Cache URL
  - URL Sig
  - URI Signing

Note this is not an exhaustive of required files, and many files
must be dynamic and cannot be inferred.

But this does remove the manual configuration for this list.
More files may be added in the future.

* Refactor ConfigGen parent.config to reduce code

Refactors parent.config generation to have a single DS for-loop,
instead of different loops for top-level vs non-top-level caches.

This is to prepare for Topologies, which will need their own
behavior, irrespective of "top level" Edge vs Mid, for every DS.

* Add ORT atstccfg Topology support

* Fix code broken from changed TO Client functions

* Fix log typo

Also adds a TODO to propogate, once config gen returns errors.

* Removed commented code

* Remove unused code, add topology name to comment

* Add ORT Topology Header Rewrites

* Add ORT atstccfg error log for parentless topology

* Add ORT atstccfg new req for update_status

Necessary because the latest API considers Topologies, previous API
doesn't.

* Fix tests for atscfg topology/capability change

* Change ORT/atstccfg to break outer loop

Optimization.

* Change ORT atscfg variable name

Per code review.

* Change ORT atscfg to use topology map

More performant for large numbers of topologies. Per PR review.