Adds support for CSV parsing in astats (#4993)
* Adds support for CSV parsing in astats

The astats parser will check the incoming content-type header to determine if it should parse the data as json or csv. Added a cfg option of http_polling_format that defaults to "text/json", this is what is sent on http requests to caches. Changing it to "text/csv" will enable the csv output if the cache's astats plugin as has support for CSV, otherwise astats will respond with "text/json" or "text/javascript" depending on the version.

* Fix - remove unused, RHV slices, err msg capitalization, import groupings

* Use loop for delete, fix err msg, remove log of output vars since we just care that parsing succeeded

* Change to do more allocation up front

Now scans in all data in to a string slice which then lets us allocate the data map up front.  In benches this bumps the performance up from ~8000 to 10000