Add csv support to stats_over_http (#5023)
* Add csv support to stats_over_http

This uses the same parsing method as astats csv and provides almost double the performance with less than half the allocations of the json version.

Also add a null check for dsStats. Its possible to do a request to ATS to a DS that does not exist, ATS will then insert stats for that request into the remap stats list. When it gets to TM though it will not be in the list of known DSs so if we try to index by it it will end up null

* * Update changelog to mention ATS 9
* Remove my dstat nil check since this has been merged in already

* Include information about users needing system_stats.so for SoH

* Added information about including the system_stats plugin when using SoH and that it can be used to provide system info for any custom extensions