Fix ORT/atstccfg to use v1-client not vendor
This fixes a breaking issue with the vendored client returning
tc.ServerNullable, which was changed to be a V2 object not v1.

Because of the member IPIsService which doesn't exist in V1 and
defaults to false when it doesn't exist, this breaks config gen
because all interfaces are "not service" and thus not used.

This fixes the bug by using v1-client which returns a ServerV1
which ORT/atstccfg then upgrades properly, forcibly setting
all IPIsService to true.