Change ORT/atstccfg to use standard TC objects (#5247)
* Change ORT/atstccfg to use standard TC objects

* Remove unused symbols

* Change atscfg makeCGMap to unexported

* Change atscfg parent symbol to unexported

* Change atscfg isTopLevelCache to unexported

* Change atscfg getATSVersion to unexported

* Change atscfg genericProfileConfig to unexported

* Remove atscfg unused func

* Change atscfg makeTopologyNameMap to unexported

* Change atscfg filterDSS to unexported

* Change atscfg filterParams, toMap to unexported

* Change atscfg getServerIPAddress to unexported

* Change atscfg cacheConfig symbols to unexported

* Change atscfg cacheurl symbols to unexported

* Change atscfg chkconfig symbols to unexported

* Change atscfg headerRw internals to unexported

* Change atscfg ipallow internals to unexported

* Change atscfg meta internals to unexported

* Change atscfg package internals to unexported

* Change atscfg parent internals to unexported

* Change atscfg regexremap internals to unexported

* Change atscfg regexreval internals to unexported

* Change atscfg remap internals to unexported

* Fix atscfg package test

* Change atscfg unknown internals to unexported

* Change atscfg sslmulticert internals to unexported

* Change atscfg topohdrrw internals to unexported

* Fix atscfg for tc/client changed server

* Change atscfg to DS alias for stability

Changes lib/go-atscfg and traffic_ops_ort/atstccfg to use an alias
for DeliveryService, so all of atscfg doesn't have to change every
time the Traffic Ops Client type is changed, but only the alias
and any broken symbols.

* Fix ORT/atstccfg to use v1-client not vendor

This fixes a breaking issue with the vendored client returning
tc.ServerNullable, which was changed to be a V2 object not v1.

Because of the member IPIsService which doesn't exist in V1 and
defaults to false when it doesn't exist, this breaks config gen
because all interfaces are "not service" and thus not used.

This fixes the bug by using v1-client which returns a ServerV1
which ORT/atstccfg then upgrades properly, forcibly setting
all IPIsService to true.

* Add atscfg warnings for nil values

* Fix atscfg godoc

(cherry picked from commit 81c9fd0cc283d7f8c6166827a4f1367cb1efec1a)