Added transitive source build for ATS. (#5004)
* Added transitive source build for ATS.

This provides a basic Apache Traffic Server compile, complete with relevant
libraries, that can be used along with Traffic Control.

* Fixed line endings on two lines of pkg.

The lines had spaces where the file intended to use tabs. This has been
corrected.

* Removed stray "s from help lines.

These were introduced erroneously as the result of a merge conflict resolution.

* Re-added erroneously removed " from pkg help documentation.

* Added comment describing which specific version of tsb is being pulled.

* Fixed potential error in build where temporary dist dir was not cleaned up.

It was being left in place when the build did not succeed.

* Removed the update of permissions in the dist directory.

This only works when pkg is run with enough permissions to modify the files it
creates.