Show correct CDN ID in Changelog for snapshots (#5195)
Currently, when a CDN snapshot has taken place a log entry with the
CDN ID is written to the Changelog. The handler is still responsible
for maintaining backwards compatibility with v1 API parameters.

As such, the log entry is currently only logging the CDN ID should
the API parameters match v1 legacy parameters.

Going forward the CDN ID shown in the changelog will match either
the v1 legacy "id" parameter or the newer "cdnID" parameter.