Ensure SSL changes are logged.
There are three endpoints (Add, Generate, Delete) that manipulate SSL certificates
for a delivery service. (Actually there are more with the Let's Encrypt ACME
endpoints, but those have a different changelog order of operations). These all
must log their action in the Changelog for verification and confirmation when
the actions complete. Generate and Delete succesfully log the changes, however
Add was apparently not.

In fact, there are two successful cases where the SSL keys could be added and
the endpoint would return prematurely, preventing the action from being logged
in the Changelog. The Changelog entry is now performed before the return of
these two logic flows.

Additionally added a comment to a public package function (Generate) and clarified
language within the Changelog messages.