Prevent tenant update to chose child as new parent
As it stands currently, there is a bug where you can update a Tenant
and select a child of said tenant to be the new parent.
This causes a circular reference in the database with the adjanceny
list and then prevents the child and parent from being returned in the
recursive CTE, which makes it appear as though those values have been
deleted.
> Note that should this happen it is still possible to salvage the data
by manually editing the `tenants` table and fixing the incorrect
parent/child relationship.

To prevent the problem, a check is done to ensure the chosen ParentID
does not reside in the current list of children when attempting to
update the Tenant.