Fix potential steering registry race, improve steering logging
Clearing the existing map and putting all the new steerings into it can
potentially cause a race where the map is empty or partially empty when
doing steering lookups for incoming requests. Instead, swap the old map
with the new map (which is atomic).

Additionally, improve the steering registry logging so that only changes
are logged, rather than re-printing the entire registry when a single
entry is changed.