Set Traffic Router to only accept TLSv1.1 and TLSv1.2 (#5547)
The reason I did not turn off TLSv1.1 is because I had some issues getting thing to work correctly with it off. The reason I did not turn on TLSv1.3 is because it is not supported in CentOS 7.

TLSv1 and TLSv1.1 have been deprecated by all the major browsers since March of 2020. We might want to look at logging the negotiated protocol if that hasn't already been done.

(cherry picked from commit 9718699496098978ad615b20a1ad55dedee777e4)