Update ORT/atstccfg old client to V3
This updates the old client to V3.

The new client is still V3, because the v4-client isn't done yet. It
adds a TODO comment, so when v4 is ready, ORT should be changed to
use it.

It also refactors the client usage a bit, so the toreq user
doesn't have to check and fall back itself, but rather the entire
client falls back. Making it easier to use and more efficient.