Only use monitored interfaces for vitals
Per bug report #5695, any and all interfaces returned in polling
were used to calculate vitals, such as bytes in or out, bandwidth
and more. However, only designated interfaces should be used for
calculating such metrics and vitals. As such, this commit will
check for monitored interfaces and only use those for the calculations.