Add -M option to specify a url for a maxmind database download
t3c first checks if a url exists, then it grabs the timestamp on the existing gz file (if it exists).
Then issues an IMS with the date to determine if it needs to update the db, if so it downloads and
extracts the db to the ats config dir.