Fix t3c-diff to return diff for  nonexistent file
It's important that t3c-diff always return a diff if either file
doesn't exist.

Without this, new semantically empty files won't be created.

It's imperative that empty files be created. For example, if they are
referenced by a plugin, it may be legal for them to be empty,
but ATS will fail to load if they don't exist.