Fix t3c unsetting to update flag (#5875)
* Fix t3c to unset traffic ops update flag

This was fixed before, it seems the t3c-apply PR had a bad merge and
undid that. This is just re-applying the same code.

* Add t3c test update flag is unset after syncds

Also fixes broken git tests that were nondeterministically failing,
because git was deleted from the wrong directory and tests unrelated
to git were creating git repos.