Fix t3c reload (#5873)
* Fix t3c not reloading

Bug was the data of what files and packages were changed not being
set. This is a critical part of the t3c-reload call, Guessing it
disappearing was a merge failure from the reload PR being so old.

* Add t3c reload test

* Changed t3c test fixture ATS version to "CHANGEME"

Also changes the baseline-config as necessary. I confirmed the change
is correct, the previous was being caused by the ATS version being
detected as 5, confirmed in the test log as:
"Unsupport version of ats found 5"
from lib/go-atscfg/remapdotconfig.go