Fix t3c reval to reload (#5887)
Reval reload was failing because reval doesn't get and check packages
from Traffic Ops. Which it doesn't need to, but the func to check
if a package was installed was only checking the cache populated by
the TO fetch.

This fixes it to check rpm if the requested package isn't in the
cache, which then fixes the call to know trafficserver is installed,
which then fixes the thing that calls that to reload ATS correctly.