Add define switch for ATS version (#5849)
From versions 8->9->master the api definitions for the TSCont* functions have been changing. From 8 to 9 TSContSchedule dropped an argument, and 9 to master the function
was dropped completely in favor of TSContScheduleOnPool which took the original amount of arguments.

Adding a switch to for <9 use the old definition and for others use the TSContScheduleOnPool function which should act the same. This way it will always build (for now)