Add atscfg storage.config generation tests
Also fixes a poorly behaving failsafe. If a drive prefix parameter
exists but no drive letters, it should warn but not insert it.
It was inserting a malformed prefix with no letter, e.g. '/dev/sd'.
This fixes it to not insert, in the event someone's Parameters are
malformed and have a prefix but no letters.