Fix TV Postgres reencrypt tool
The "id" column of the "sslkey" table was removed, but the queries in
the reencrypt tool were not updated. Change the queries to select and
update by the "deliveryservice" (XML ID) column instead.

Also, fix some error logging where the error was not being appended to
the message.

Closes: #6061

Add tests, fix some permissions issues with the TV tools

Add README.md to explain the tests

Update the TO DB GitHub Actions tests workflow to include TV DB