Caching efficiency improvements
Using overlay2 as a driver, this reduces space used in the virtual fs
from 686MB to 611MB, and reduces the image cache size from 1.7MB to
1.5MB, for a total disk space saving of about 75MB.

It does, however, increase build time from about a minute and a half to
about two minutes when building from scratch (my machine). When the RPM
changes but steps before that are cached, the build time changes from 22
seconds to 36 seconds, so for the average use-case the build time change
is about +1.75x/+14 seconds (my machine).