Take topology-based DSes into account when checking DSS assignments (#6408)
* Take topology-based DSes into account when checking DSS assignments

In order to prevent accidental ouages, TO currently has validation which
prevents things like deleting the last server that is assigned to an
active delivery service. However, this validation ignored the fact that
delivery services can now be assigned to a topology instead of using DSS
assignments. This prevents a number of actually valid things, such as
safely setting caches to ADMIN_DOWN, assigning MSO servers to
topology-based DSes, and removing legacy DSS EDGE assignments from
topology-based DSes. This fixes that validation in order to take
topology-based DSes into account.

Closes: #6392

* Fix GROUP BY

* Add TO API test