Correcting use of httpsPort instead of secure.port (#6538)
* Correcting use of httpsPort instead of secure.port

I was trying to change the TrafficRouter SecurePort value to other than 443, and tried to use the parameter secure.port added to the server.xml on the CCR profile in Traffic Portal. But noticed, that TcOps/TcMon, do not add that parameter to the CrCronfig.json.I digged into the code and fount that the use of secure.port was added to TrafficRouter on the following commit: https://github.com/apache/trafficcontrol/commit/8ff9006beb0639a21ac3235cdeaa15bf852cfc1a However, no counter part was added to TcOps to actually state the value secure.port from the TrafficPortal paramenters. The value that TcOps adds to the CrConfig.json is the httpsPort, from the server configuration itself on TrafficPortal, but that parameter is not used anyowhere by Traffic Router (from what I saw).I see that it would require just two lines edits on traffic_router/core/src/main/java/org/apache/traffic_control/traffic_router/core/util/LanguidState.java to make use of httpsPort and not the unexisting secure.port.Basically, edit lines 86 and 87 of the file, and replace secure.port for httpsPort

* Correcting use of httpsPort instead of secure.port

* correcting to use hasNonNull instead of has