Add t3c final failure log msg (#6586)
This adds a final obvious failure or success message to t3c. It isn't
obvious whether error logs are fatal, or if t3c was able to keep
going. This makes it obvious whether it succeeded or failed,
and makes it easy to see, to search logs, and to create alarms
when the run fails.

Also adds panic catching and logging.