Fix t3c Capability to not continue on error
This bug causes catastrophic config gen errors if the TO request
for ds_required_caps fails but server_reqs succeeds,
because then no servers have required caps, causing remap lines
for all DSes with required caps to be omitted.

This fixes it by making capability failures fail, rather than
continuing.

It formerly continued, because when capabilities were introduced,
it was necessary to continue if TO was older than t3c/ORT. But
now, the fallback API major version is 3.0, which has Capabilities.

If this bug wasn't fixed by simply always failing, it would have to
be fixed by making a server_caps error also set ds_req_caps to nil.