Reduce TM dependency on CRConfig (#6625)
* Reduce TM dependency on CRConfig

Do not use CRConfig data when the same data already exists in the
TMConfig. For data that TM needs but doesn't exist in TMConfig, add it
to TMConfig. This way, TM only uses data in the TMConfig and can become
simply a shuttle for the CRConfig without actually using any CRConfig
data.

Closes: #6512
Related: #1739
Related: #5367

* Make trafficServer deliveryService field backwards-compatible

* Address review comments

* Fix grave (accent) mistakes, clarify field omission