t3c race condition time stamp updates
Remove the ability to set "update" times via the
/servers/:hostname/update endpoint (only "apply" times are writable via
this endpoint).

Fix docs to reflect the existence of the "updated" and "reval_updated"
query paramters in the /servers/:hostname/update endpoint (kept for
backwards-compatibility).

Correct the date on the DB migration so that it applies after migrations
that were merged earlier.