First commit for fixing issue with traffic_router log rotation
Added new file for log4j2.xml as well as logrotate configuration. We may need to tweak further, maybe add an hourly logrotate file if we feel like. But this is a good start. I need to know the other places that I need to modify Traffic Router to make sure the changes get into the RPM.