TR: use hostname from CRConfig if only a single TR in CRConfig (#7205)
In situations where there is only a single TR in the CRConfig but
actually multiple TRs in the system, using the server hostname in zone
generation leads to the creation of invalid mname (master name) field in
the SOA record unless the server hostname actually matches the single
hostname in the CRConfig. In this scenario, it's preferable to just
assume the hostname of the single TR in the CRConfig so that the
resulting SOA mname is a real, resolvable name. That way all the TRs in
the system will use the same SOA mname.

Otherwise, TR servers that aren't in the CRConfig will generate an SOA
mname like my-tr.mycdn.example.com without actually resolving that to an
A/AAAA record.