Handle log rotation using logrotate (#7234)
* First commit for fixing issue with traffic_router log rotation

Added new file for log4j2.xml as well as logrotate configuration. We may need to tweak further, maybe add an hourly logrotate file if we feel like. But this is a good start. I need to know the other places that I need to modify Traffic Router to make sure the changes get into the RPM.

* Changes to build.sh and pom.xml files for building of TR

Addition of logrotate. Still need to test.

* Update CHANGELOG.md

* Update traffic_router.rst

Co-authored-by: Villa, Joseph <Joseph_Villa@comcast.com>